fix: group에 멤버 수 컬럼 추가하여 최적화 진행

+++ app/api/groups/[groupId]/route.ts
@@ -10,14 +10,14 @@ export async function GET(
     const uuid = await getSupabaseUuid();
 
     if (!uuid) {
-      return new NextResponse("User not found", { status: 404 });
+      return new NextResponse("Unauthorized", { status: 401 });
     }
 
     // 그룹 정보 조회
     const groupArr = await supabaseDb.select("groups", { id: params.groupId });
     const group = groupArr[0];
     if (!group) {
-      return new NextResponse("Not found", { status: 404 });
+      return new NextResponse("Group not found", { status: 404 });
     }
 
     // 그룹 멤버 role 조회
@@ -25,7 +25,10 @@ export async function GET(
       group_id: params.groupId,
       user_id: uuid,
     });
-    const role = memberArr[0]?.role || null;
+    if (!memberArr) {
+      return new NextResponse("Not member of this group", { status: 404 });
+    }
+    const role = memberArr[0].role;
 
     return NextResponse.json({ ...group, role });
   } catch (error) {
@@ -52,6 +55,16 @@ export async function PATCH(
       return new NextResponse("Name is required", { status: 400 });
     }
 
+    // 그룹 멤버 role 조회
+    const memberArr = await supabaseDb.select("group_members", {
+      group_id: params.groupId,
+      user_id: uuid,
+    });
+    if (!memberArr) {
+      return new NextResponse("Not member of this group", { status: 404 });
+    }
+    const role = memberArr[0].role;
+
     const group = await supabaseDb.update(
       "groups",
       {
@@ -62,7 +75,7 @@ export async function PATCH(
       { id: params.groupId }
     );
 
-    return NextResponse.json(group);
+    return NextResponse.json({ ...group, role });
   } catch (error) {
     console.error("[GROUP_PATCH]", error);
     return new NextResponse("Internal error", { status: 500 });
@@ -77,7 +90,16 @@ export async function DELETE(
     const uuid = await getSupabaseUuid();
 
     if (!uuid) {
-      return new NextResponse("User not found", { status: 404 });
+      return new NextResponse("Unauthorized", { status: 401 });
+    }
+
+    // 그룹 멤버 role 조회
+    const memberArr = await supabaseDb.select("group_members", {
+      group_id: params.groupId,
+      user_id: uuid,
+    });
+    if (!memberArr) {
+      return new NextResponse("Not member of this group", { status: 404 });
     }
 
     const group = await supabaseDb.delete("groups", { id: params.groupId });
